PmxDressup_v1.0.0

+++ setup_clear.py
@@ -2,8 +2,9 @@ import os
 import shutil
 from glob import glob
 
-shutil.rmtree("build")
+if os.path.exists("build"):
+    shutil.rmtree("build")
 
-for source in glob("mlib\\**\\*.pyd", recursive=True):
+for source in glob("mmd_base/**/*.pyd", recursive=True):
     print(f"remove {source}")
     os.remove(source)
